Changed device index type to size_t from int
Device index type changed to size_t in master branch.
So changed all relevant code with size_t.

+++ service/resource-directory/client/src/rd_client.c
@@ -59,7 +59,7 @@ _add_resource_payload(CborEncoder *parent, oc_resource_t *resource, char *rel,
 
 #ifndef ST_APP_OPTIMIZATION
 bool
-rd_publish(oc_endpoint_t *endpoint, oc_link_t *links, int device_index,
+rd_publish(oc_endpoint_t *endpoint, oc_link_t *links, size_t device_index,
            oc_response_handler_t handler, oc_qos_t qos, void *user_data)
 {
   char uuid[OC_UUID_LEN] = { 0 };
@@ -139,7 +139,7 @@ rd_publish_with_device_id(oc_endpoint_t *endpoint, oc_link_t *links,
 #endif /* ST_APP_OPTIMIZATION */
 
 bool
-rd_publish_all(oc_endpoint_t *endpoint, int device_index,
+rd_publish_all(oc_endpoint_t *endpoint, size_t device_index,
                oc_response_handler_t handler, oc_qos_t qos, void *user_data)
 {
   if (!endpoint || !handler) {
@@ -186,7 +186,7 @@ rd_publish_all(oc_endpoint_t *endpoint, int device_index,
 
 #ifndef ST_APP_OPTIMIZATION
 bool
-rd_delete(oc_endpoint_t *endpoint, oc_link_t *links, int device_index,
+rd_delete(oc_endpoint_t *endpoint, oc_link_t *links, size_t device_index,
           oc_response_handler_t handler, oc_qos_t qos, void *user_data)
 {
   char uuid[OC_UUID_LEN] = { 0 };
